1290. Convert Binary Number in a Linked List to Integer Difficulty - Easy Tag - Linked List
Time Complexity = O(n)
Space Complexity = O(1)

Given head which is a reference node to a singly-linked list. The value of each node in the linked list is either 0 or 1. The linked list holds the binary representation of a number.

Return the decimal value of the number in the linked list.

Signed-off-by : Akib <akibislamr5@gmail.com>