237. Delete Node in a Linked List
Difficulty - Easy
Tag - Linked List

Time Complexity  = O(1)
Space Complexity = O(1)

Write a function to delete a node in a singly-linked list. You will not be given access to the head of the list, instead you will be given access to the node to be deleted directly.

It is guaranteed that the node to be deleted is not a tail node in the list.

Signed-off-by : Akib <akibislamr5@gmail.com>